==> Code/1_User_Reviews_Scraping.ipynb <==
!pip install google_play_scraper

from google_play_scraper import reviews_all, app, Sort
import pandas as pd
import numpy as np

app_id_to_name = {
    'com.kai.kaiticketing': 'Access by KAI',
    'com.mrt.jakarta': 'MyMRTJ',
    'com.jaklingkoindonesia.app': 'Jak Lingko App',
    'com.transjakmobile': 'TJ: Transjakarta'
}

# List of app IDs to scrape reviews from
app_ids = list(app_id_to_name.keys())

# List to hold all reviews
all_reviews = []

# Iterate through each app ID and get reviews
for app_id in app_ids:
    result, continuation_token = reviews(
        app_id,
        lang='id',
        country='id',
        sort=Sort.NEWEST,
        count=9000,
        filter_score_with=None
    )

    # Append the app name to each review in the result
    for review in result:
        review['app_name'] = app_id_to_name[app_id]

    all_reviews.extend(result)

df_review = pd.DataFrame(np.array(all_reviews),columns=['review'])
df_review = df_review.join(pd.DataFrame(df_review.pop('review').tolist()))

df_review['at'] = pd.to_datetime(df_review['at'], errors='coerce')

my_df = df_review[['score','at', 'content', 'app_name', 'appVersion']]

my_df.to_csv("GooglePlay.csv", index = False)
