==> Code/5_Save_to_HuggingFace.ipynb <==
from huggingface_hub import login

login(token="xxxx")

from huggingface_hub import whoami

whoami()

from huggingface_hub import HfApi

api = HfApi()

!ls /content/drive/MyDrive/Bach_Thesis/Models/SA_Check/indobenchmark_indobert-base-p1/rosncl-check-1/Run_6/model

config.json	   special_tokens_map.json  vocab.txt
model.safetensors  tokenizer_config.json

api.upload_folder(
    folder_path="/content/drive/MyDrive/Bach_Thesis/Models/SA_Check/indobenchmark_indobert-base-p1/rosncl-check-1/Run_6/model",
    repo_id="chimons-academy/indobert-jkt-transpub-app-review",
    repo_type="model"
)

api.upload_folder(
    folder_path="/content/drive/MyDrive/Thesis_NS/Models/TM_Tuned_2/Config/positive/bertopic",
    repo_id="chimons-academy/bertopic-jkt-transpub-app-pos-review",
    repo_type="model"
)

api.upload_folder(
    folder_path="/content/drive/MyDrive/Thesis_NS/Models/TM_Tuned_2/Config/negative/bertopic",
    repo_id="chimons-academy/bertopic-jkt-transpub-app-neg-review",
    repo_type="model"
)
